# Checkout concurrently (~2.5x speedup)