--- a/Source/CarthageKit/ProducerQueue.swift
+++ b/Source/CarthageKit/ProducerQueue.swift
@@ -8,7 +8,7 @@
 
 import ReactiveSwift
 
-/// Serializes or parallelizes (up to a limit) the execution of many 
+/// Serializes or parallelizes (up to a limit) the execution of many
 /// SignalProducers, like flatten(.concat) or flatten(.merge), but without all
 /// needing to be enqueued in the same context.
 ///
@@ -16,55 +16,60 @@ import ReactiveSwift
 /// to the queue object, instead of being required to funnel all producers
 /// through a single producer-of-producers.
 internal final class ProducerQueue {
-	private let concurrentQueue: DispatchQueue
-	private let serialQueue: DispatchQueue
-	private let semaphore: DispatchSemaphore
+    private let operationQueue: OperationQueue
 
 	/// Initializes a queue with the given debug name and a limit indicating the
 	/// maximum number of producers that can be executing concurrently.
-	init(name: String, limit: Int = 1) {
-		concurrentQueue = DispatchQueue(label: name.appending(".concurrent"), qos: .userInitiated, attributes: .concurrent)
-		serialQueue = DispatchQueue(label: name.appending(".serial"), qos: .default)
-		semaphore = DispatchSemaphore(value: limit)
-	}
+    init(name: String, limit: Int = 1) {
+        operationQueue = OperationQueue()
+		operationQueue.name = name
+        operationQueue.maxConcurrentOperationCount = limit
+    }
 
 	/// Creates a SignalProducer that will enqueue the given producer when 
 	/// started.
 	func enqueue<T, Error>(_ producer: SignalProducer<T, Error>) -> SignalProducer<T, Error> {
 		return SignalProducer { observer, disposable in
-			self.serialQueue.async {
+			let operation = ManuallyFinishingOperation { operation in
 				if disposable.isDisposed {
+					operation._isFinished = true
 					return
 				}
 
-				// Prevent more than the limit of operations from occurring
-				// concurrently.
-				//
-				// Block the serial queue to prevent creating a new thread on 
-				// the concurrent queue just to immediately block it.
-				self.semaphore.wait()
+				producer.startWithSignal { signal, signalDisposable in
+					disposable.add(signalDisposable)
 
-				self.concurrentQueue.async {
-					if disposable.isDisposed {
-						self.semaphore.signal()
-						return
-					}
+					signal.observe { event in
+						observer.action(event)
 
-					producer.startWithSignal { signal, signalDisposable in
-						disposable.add(signalDisposable)
-
-						signal.observe { event in
-							observer.action(event)
-
-							if event.isTerminating {
-								self.semaphore.signal()
-							}
+						if event.isTerminating {
+							operation._isFinished = true
 						}
 					}
 				}
 			}
+
+			self.operationQueue.addOperation(operation)
 		}
 	}
+}
+
+/// An block operation that can only be finish by setting its _isFinished
+/// property to true.
+fileprivate final class ManuallyFinishingOperation: BlockOperation {
+    override var isFinished: Bool {
+        return _isFinished && super.isFinished
+    }
+
+    var _isFinished: Bool = false {
+        willSet { willChangeValue(forKey: "isFinished") }
+        didSet { didChangeValue(forKey: "isFinished") }
+    }
+
+    init(_ block: @escaping (ManuallyFinishingOperation) -> Void) {
+        super.init()
+        addExecutionBlock { block(self) }
+    }
 }
 
 extension SignalProducerProtocol {