With our test project with 26 checkouts, before:

```
$ time carthage checkout
carthage checkout  11.63s user 5.81s system 97% cpu 17.848 total
```

after:

```
$ time carthage checkout
carthage checkout  13.68s user 6.97s system 270% cpu 7.618 total
```