@erichoracek 2017-04-29
Per the discussion in #1839