@erichoracek 2017-04-29
After taking a look at bootstrapping with a cold checkouts cache (removing `~Library/Caches/org.carthage.CarthageKit` before bootstrap), this should also be sped up by 2x with the latest commit.

before:

```
$ carthage checkout
carthage checkout  38.90s user 18.98s system 99% cpu 58.040 total
```

after:

```
$ carthage checkout
carthage checkout  36.84s user 16.64s system 40% cpu 2:12.18 total
``` 

As far as I can tell, there's no reason that cloning to the same directory would need to be on the same scheduler (although I do understand why adding submodules would be).