@mdiep 2017-04-30
I feel a little uncomfortable with this until we can upgrade ReactiveSwift and get the new `.concurrent` merge strategy to limit the number of concurrent operations. 😕