@andersio 2017-04-30
`OperationQueue` can be used as a transitional solution.