@erichoracek 2017-05-01
I've updated this logic to checkout concurrently with a limit of the active number of processors. The speed seems unaffected, but it should hopefully prevent unbounded thread growth with a large dependency graph.