@andersio 2017-05-01
Is there a problem with using an `OperationQueue` and `maxConcurrentOperationCount`? While both dispatch queues are at the same QoS classes in https://github.com/Carthage/Carthage/pull/1910/commits/58237f004848254b0a38a3842e11962a6ba80806, I'd still rather avoid semaphores, or at least the concurrent queue should get higher priority.