@erichoracek 2017-05-01
The primary issue in my mind was that migrating to use `OperationQueue` would require a decent amount of code around managing an `Operation` subclass that wrapped a `SignalProducer` (since `BlockOperation` would be insufficient by itself as far as I could tell. It seemed like this may not be worth implementing when we'll soon have a first-class `.concurrent(...)` flatten strategy.

Can you briefly explain why we'd want to ensure that the concurrent queue has a higher priority?