@andersio 2017-05-01
ICYMI you can use `OperationQueue` like a `DispatchQueue` with `BlockOperation`.

> Can you briefly explain why we'd want to ensure that the concurrent queue has a higher priority?

Semaphore is unsafe due to priority inversion. After a second thought, it appears that marking the concurrent queue is not enough, if the `producer` starts the work that could yield terminal event asynchronously in lower QoS classes.