@erichoracek 2017-05-01
Makes sense—I understand why this would have priority inversion issues from the perspective of the queues having different priorities, but what is the issue if they share the same priority?