@erichoracek 2017-05-01
With respect to using `BlockOperation` and `OperationQueue`, are you recommending that concurrent signal producers are implemented by blocking the `BlockOperation` using a construct like `SignalProducer.wait()` within the `BlockOperation`'s block?