@andersio 2017-05-01
Hmm, Sorry about overlooking the needs here. [An `Operation` subclass is probably necessary](https://gist.github.com/andersio/33faf2fb9eddefee9bab80c1ea34ccda). Blocking operators use semaphore, so they won't help here.

> but what is the issue if they share the same priority?

Within the scope of this PR, there isn't an issue. But unless we are sure that every producer started by `ProducerQueue` performs work and sends events on the same QoS class as the `ProducerQueue`, it cannot be declared priority inversion safe.