@erichoracek 2017-05-02
I've migrated to `OperationQueue` using your suggested `BlockOperation` subclass @andersio—please take a look when you have a chance. Thanks!